decimal issue fix

+++ src/main/java/com/asg/finance/dto/BankPaymentChargeDetailRequest.java
@@ -1,5 +1,6 @@
 package com.asg.finance.dto;
 
+import java.math.BigDecimal;
 import lombok.Data;
 
 @Data
@@ -7,14 +8,14 @@ public class BankPaymentChargeDetailRequest {
 
     private Long detRowId;
     private Long chargePoid;
-    private Long chargeAmount;
+    private BigDecimal chargeAmount;
     private String description;
     private String remarks;
     private String refDocId;
     private Long refDocPoid;
     private Long fdaDetRowId;
     private Boolean selected;
-    private Long pdaAmount;
-    private Long ffAmount;
+    private BigDecimal pdaAmount;
+    private BigDecimal ffAmount;
     private String actionType;          // "isCreated", "isUpdated", "isDeleted", "noChanges"
 }
